requirements.txt
# KISA TTPs Pipeline Dependencies

# PDF Processing
pdfplumber==0.11.0
PyMuPDF==1.24.0
Pillow==10.2.0

# AI/LLM
anthropic>=0.40.0

# Data Processing
pyyaml==6.0.1
jinja2==3.1.3

# MITRE ATT&CK
mitreattack-python==3.0.6

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Visualization
graphviz>=0.20.0  # Kill Chain 시각화
